requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-js-asset==2.0.0
sqlparse==0.4.2
